requirements.txt
streamlit
seleniumbase
webdriver-manager
pandas

